Final touch for release of TriPaint 1.2
