+++ .travis.yml
@@ -1,6 +1,3 @@
-# Use container-based infrastructure
-sudo: false
-
 # These directories are cached to S3 at the end of the build
 cache:
   directories:
@@ -14,6 +11,6 @@ before_cache:
 
 language: scala
 scala:
-  - 2.12.6
+  - 2.12.11
 script:
   - sbt coverage test coverageReport coveralls
