+++ src/main/scala/com/martomate/tripaint/view/TriPaintViewFactory.scala
@@ -1,0 +1,7 @@
+package com.martomate.tripaint.view
+
+import com.martomate.tripaint.model.TriPaintModel
+
+trait TriPaintViewFactory {
+  def createView(controls: TriPaintViewListener, model: TriPaintModel): TriPaintView
+}
